rework project structure; introduce JavaFX Frontend; update dependencies
- resolves #20
- introduction of new Frontend with JavaFX (WIP) -- contributes #9
