--- Backend/src/main/java/de/metallist/backend/BackendApplication.java
@@ -1,13 +1,0 @@
-package de.metallist.backend;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class BackendApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(BackendApplication.class, args);
-    }
-
-}
