+++ Backend/src/main/resources/application-dev.properties
@@ -7,6 +7,8 @@ management.endpoints.web.exposure.include=*
 
 # logging
 logging.level.root=DEBUG
+logging.level.org.springframework.boot=DEBUG
+logging.level.web=DEBUG
 logging.file.name=./logs/Contractcollection.log
 logging.logback.rollingpolicy.file-name-pattern=./logs/Contractcollection-%d{yyy-MM-dd}.%i.log
 logging.logback.rollingpolicy.max-history=10
