+++ Backend/src/main/resources/static/contractcollection-api.yaml
@@ -10,6 +10,15 @@ info:
 servers:
   - url: 'http://localhost:8080'
 paths:
+  /status:
+    get:
+      summary: serves as status checker
+      operationId: getStatus
+      responses:
+        '200':
+          description: backend is running
+        '404':
+          description: backend is not running or can't be accessed
   /add:
     post:
       summary: creates a new contract
