+++ Backend/src/test/resources/testExport.txt
@@ -1,1 +1,1 @@
-[0, 0, 0, 12, 43, 94, 8, 54, 108, -21, -38, -68, 89, 91, 127, 91, -52, -12, -122, 38, -47, 35, -117, -122, -3, 97, -81, -7, 102, -32, 16, 62, -100, -73, 114, -62, -33, 10, -126, 65, 29, 55, -6, 8, 45, -16, 89, -29, 61, 28, -48, -84, -4, 69, 8, 41, -76, 47, -28, 45, 54, 60, -37, 84, -68, -15, 66, 56, 11, 31, -8, 107, 79, -27, 67, -44, 77, 39, -11, 93, -115, -97, -70, -30, -53, 126, -92, 49, 40, -43, -128, 46, 18, 97, 3, 58, 88, 101, 4, 76, -85, 3, 80, -33, 91, 26, -15, 72, 96, 27, 44, 105, -33, 80, -77, 114, -49, -66, 4, 93, 3, -17, -66, 38, 44, -35, 105, 7, -32, 118, 87, 101, -26, 13, 65, 48, 24, 111, 6, 93, 25, 90, -19, 39, 73, 27, 55, -25, 71, -121, -112, -70, -118, 8, -66, -5, 31, -85, 35, -104, -34, -78, 114, -92, -49, 78, 32, 102, 7, -122, -106, 126, -113, 15, 82, 61, 85, 116, -32, 95, -84, 10, 77, -86, -43, -13, -101, 4, 58, 46, -70, -93, -51, 73, -20, 32, -67, -113, 75, -52, -99, 101, -104, 45, 3, 43, -80, 7, -37, -49, -35, -88, 90, -78, -5, 21, 8, 23, -100, 31, 113, -53, 2, -71, -112, 89, -105, 4, -38, 40, -55, -27, -64, 38, -16, -65, -78, 51, 50, 48, -59, -89, 40, 92, -64, 32, -22, 114, 39, -102, 24, 2, 57, -71, -114, -126, -22, 124, -34, -16, 115, -13, 91, -104, 90, 111, 8, -1, 25, 61, 33, 121, -58, -3, -55, 29, -45, -101, -16, 85, 87, 54, 59, -39, -92, -44, 3, 37, 85, 104, 52, -12, 75, -128, 20, -119, -15, -59, -108, -70, -39, 51, 97, -84, 76, -120, 16, -4, -5, -31, -61, -37, 82, 92, 58, -47, 56, -46, -74, 37, -73, -21, 21, 121, -119, 125, 44, 104, -37, 32, -105, -93, 107, -119, 99, 30, -91, 88, 54, 105, 27, 9, -45, -36, -88, 18, 78, 110, 88, 114, -123, -27, 115, 74, 98, 37, -56, -107, -79, -24, -62, -107, -4, -98, 125, 33, 19, 4, 83, -123, 46, -48, 21, 43, 5, -68, 23, 114, 44, -93, 72, -82, -118, 117, -124, -19, 55, -51, 125, 68, 113, 100, -92, -14, -9, 49, -39, -80, -37, 23, -108, -19, -42, -58, 63, -52, 76, 30, 7, 56, -75, -81, -107, -85, -52, 119, -64, -22, 12, 95, 112, 31, -80, 12, -69, -18, -127, -31, 114, -128, 18, 64, 36, -17, 111, 127, -104, 83, 26, -2, 55, -47, -105, -57, 92, -38, 88, -32, 16, -41, -39, 55, 5, 4, -42, -74, -15, -128, -1, -38, 29, 20, -15, -94, 124, -64, 38, -33, 43, -117, -127, 100, -24, -21, -114, -34, 93, 91, 98, 112, -4, -1, 74, -57, 69, -24, 97, -128, 58, -9, 0, -40, 91, -39, -61, -20, -4, 46, 93, 127, 25, 77, 58, -72, -53, 19, 32, -95, 83, -34, -44, -127, 87, 36, 74, -50, -78, -64, -118]
+[0, 0, 0, 12, 127, -54, 25, -90, 17, -43, -85, -50, -76, -90, -124, -118, 122, -84, -7, -126, 86, -58, 123, -93, 32, 73, 112, 74, -78, -73, 68, -48, 71, 127, 91, 76, -1, -10, -80, -10, 49, -5, -18, -90, 62, -55, 50, -66, 105, -85, 86, 48, -7, 11, 108, -96, 25, -94, 127, -62, 73, 27, 41, -70, 7, -64, 85, 75, 35, -18, -60, -37, -112, -55, -96, 66, 44, -108, 11, 71, -57, 107, 109, -29, 52, -41, -71, 75, 23, 82, -90, 90, -43, -103, 54, -32, -56, -56, -24, 72, 58, 40, 95, -81, 82, 25, 83, -21, -60, 109, 26, -15, -17, -41, -107, -37, -53, -41, 63, -19, -21, 38, 97, -110, 100, -11, 53, 50, 63, 54, 81, -72, 40, 55, -45, 115, 69, -119, 61, 49, 91, -78, -93, 120, 81, 32, -118, -36, 42, -44, 36, 104, -25, -30, 111, 59, 41, -21, -28, 67, -6, 36, 74, 9, 99, -105, -70, 3, 48, 0, 54, 85, -118, -86, -28, -40, -14, -107, 36, -52, -87, -121, -16, -74, -53, -87, -45, 61, 38, -71, 115, -17, -128, 88, 58, 51, 107, 37, 119, -110, -120, -3, -25, -115, -65, 123, 17, 70, -53, -31, -42, -111, -10, 14, 46, 73, -107, 8, -93, -126, -120, -3, -34, -8, -62, 69, 5, -31, -96, 116, -96, -69, -72, -25, -76, 120, 30, -122, -2, -118, -104, -115, -23, -128, -60, -59, -65, 92, 38, -118, -67, -92, -52, 68, 21, 32, 88, -74, 29, -121, 100, -32, -25, -37, -1, -49, 51, -99, -69, -98, 47, -93, -111, -87, 74, -9, -102, 125, -73, 64, 75, 101, -125, 35, -9, 49, 29, -99, -52, 120, -19, -83, -126, -113, -119, 11, 48, -59, -73, -78, 4, -31, -103, -59, 104, -52, -128, 123, -83, -2, 68, 22, 16, 91, 62, 68, -4, 65, -105, 93, 51, 17, -112, -120, -57, 90, -74, 61, -94, -88, -93, 10, 83, 4, -57, -85, 86, 7, 85, -6, 25, -118, 126, 66, -105, -58, -106, 1, 56, -99, 84, -16, 97, 8, 107, 32, -5, -84, 60, 101, -100, -127, 45, -79, -104, -27, -63, 65, 7, 57, -52, 64, 102, -88, 29, -9, -45, -67, -75, -26, 57, -64, -61, -6, 59, -38, 103, 56, -36, -43, 123, -116, 17, -76, -125, 55, -92, 15, -47, -109, 83, 66, -74, -48, -94, -102, -60, -112, -74, -66, -48, -24, 104, 119, -105, -9, -97, -31, 8, -68, -65, -106, 86, 77, -73, 112, 54, -104, 11, 90, 11, 94, -111, -109, -32, -9, 64, 120, 114, -115, -107, -39, 70, 77, -47, -106, -3, -9, -37, 17, -27, -31, 47, 33, -27, -27, -2, 85, 118, -87, 65, -47, 105, 34, -54, 92, -47, -103, -127, 102, 54, -44, -103, -7, -8, 2, -122, -78, 115, 49, -95, 86, 10, 52, 19, 101, 11, -92, -9, -75, -56, 99, 72, 9, 91, -12, -114, -50, 119, 25, -88, -51, 63, -85, -27, 8, 7, 38, -107, 119, 82, 110, -54, -68, 48, -53, -78, 63, 72]
