+++ Backend/src/test/resources/testExportedFile.txt
@@ -1,1 +1,1 @@
-[0, 0, 0, 12, 70, 52, -79, 99, -75, 64, 77, 61, 80, 77, 33, -37, 86, -105, -116, -70, 89, -23, -88, 37, 55, -27, -3, -128, 98, 82, 32, 109, 10, -78, 110, -87, -89, 66, 50, 117, -95, -125, 54, 116, -85, 114, -64, -122, -108, 114, -89, 61, 57, 34, -98, 33, 107, -102, -84, 93, -92, 19, -84, 13, 65, -116, 57, 20, 5, 118, -65, 14, -115, 48, -41, 92, -77, -24, -44, 94, 2, 6, -21, 23, 70, -31, -62, 45, 118, 87, 73, -94, -88, 94, 11, 50, 126, 59, 119, -69, 40, 85, 122, 34, 119, 7, -18, 63, -81, 36, -28, 59, 109, -1, 107, -97, 59, -39, -84, 107, -67, 110, -30, 57, -40, 20, -115, -37, 90, 7, -67, 90, 107, 34, -89, 66, -79, 76, 91, -89, -79, 18, 108, 90, -114, 18, -37, -53, -97, 107, -75, -37, 3, 83, -105, -127, -67, 50, -75, -47, -52, 96, 13, -63, 55, -99, 34, -85, -80, 87, -78, 47, 23, 73, -85, -67, -18, 79, -41, -41, 5, -99, -7, 49, 14, -7, 2, 25, -51, 119, -16, -64, 113, -95, -28, 41, -96, 127, 20, -29, -88, 29, -113, -128, -31, 22, -6, 22, -82, -74, -105, 31, 119, 2, -24, 109, -94, -78, -121, -32, -50, 127, -18, -30, 55, 19, -60, 6, 31, -81, -33, -16, -13, -34, 105, 112, -68, 62, -114, 114, 52, 33, -80, 123, 32, -63, 121, 49, 109, 26, 79, -45, -112, -96, 0, 26, -89, -52, 90, -68, 18, 16, -31, 67, -44, 95, 9, -53, 45, 60, -34, -72, -1, -115, -109, -55, 23, -31, 64, -105, -5, 39, 44, -32, 47, 52, -16, -82, 23, -78, -128, 101, -62, 69, 55, -128, -22, 31, 54, -52, 76, 53, 55, 4, 2, 9, -13, -25, 92, 120, -75, 41, 102, 29, -112, -70, -18, -113, -12, 102, 82, -96, -11, -24, 39, 38, 48, -64, -61, 125, -14, 70, 122, 115, 93, -18, -73, 77, -23, 9, -68, 91, -7, 33, 125, 97, 64, 37, 115, 77, -107, -48, -119, -32, -85, -47, -32, 30, -33, -66, 86, -54, -56, -50, 64, 8, -10, 68, -94, -106, 17, -16, 60, 26, -117, -10, 76, 111, 27, 12, -71, 71, 115, -75, -1, -13, 120, -109, 71, 59, -10, -19, -43, 106, 70, -44, 72, -31, 54, -3, -123, 106, -19, -80, 61, 13, -7, 118, 2, -100, 39, 115, -56, -71, -74, 96, 56, -16, -61, 69, 63, 113, -73, 22, 58, -37, 13, 42, -45, -2, 101, 61, 96, 111, -62, 69, 79, 4, 15, 89, -55, 37, -45, -2, -91, 43, -38, -88, -103, 56, 88, -103, 20, -62, -121, -99, -44, -99, -126, -2, -97, 70, -10, -40, 31, 80, 122, -110, 9, 87, -128, -37, -123, -37, -101, -62, -104, 113, -57, 38, -46, -108, 35, -115, -11, -117, -102, 31, 73, 78, -45, 106, 112, 24, 7, 125, -86, 40, 47, -97, -25, 125, -74, 88, -63, 108, -106, -1, -26, 106, 34, -120, 53, 24, 38, -33, 90, -113, 24, 54, 99, 115, -118, -77, -107]
+[0, 0, 0, 12, -40, 105, -57, 67, 71, -66, 101, -116, -53, 73, -21, 113, 64, 44, 113, -99, 65, 55, 99, 121, 19, -69, -107, 39, -16, 44, 111, -47, 82, -68, 48, -30, 88, -69, -31, 42, 63, -124, -115, 75, 46, 5, -1, -25, -51, 35, -50, -66, -76, -86, 11, 82, -91, 28, -70, -106, -21, 44, 63, 80, -31, -45, -5, -45, 46, -83, -89, 83, 10, 33, -60, 21, -24, 84, -30, -13, -72, -42, -111, 32, -5, 51, 33, -13, 123, 93, -108, -17, -94, 105, -23, -128, -53, -93, -64, -20, 0, 98, -67, -119, -33, -93, -93, -1, 72, 17, -79, 105, -25, -80, 14, -126, 41, 115, 3, 53, -109, -42, 53, 17, -1, 78, 17, -51, -104, -113, -95, -111, -9, -57, -7, -108, 61, 115, 79, 16, 56, -76, 82, 13, -102, -124, -51, 84, -38, 86, -56, -33, 80, -125, 98, -57, -61, -44, 95, -73, -111, -58, -51, -24, -48, 83, 94, 67, 97, 15, -107, -32, -117, -43, 1, -88, -111, -2, 62, 0, 118, 121, -26, 72, -120, -53, 95, 55, -127, -9, 46, -110, -72, -84, 42, -78, 123, -37, -63, 111, -46, 77, 2, 38, -67, -19, 49, -94, -10, -43, 72, -115, -53, -123, 26, 113, -120, 62, 48, 65, -3, -20, -35, -79, 72, -3, -50, 93, -73, -48, -67, -102, -48, -3, 67, -69, -127, 115, -30, 6, 78, 26, -47, -16, 6, 124, -1, -64, -90, 34, 74, -106, -81, -25, -17, -2, 68, 106, 75, -81, 22, 102, 111, -117, 96, -23, 4, -53, 15, -91, 69, 22, 31, -112, -87, 92, 20, -21, -20, 17, 109, 52, -97, 53, 87, 116, -126, 91, 3, -37, -86, 24, 58, 27, 124, -3, -66, -18, -6, -128, -59, -128, -57, -54, 56, 100, -66, -40, -94, -19, 89, 1, -100, -38, -116, -109, -125, 52, 95, 26, 57, -12, -96, 23, -10, 121, 75, -73, -82, 26, -122, -110, -51, 48, 1, 113, -17, 20, -118, 103, 96, 103, 92, 11, 116, 34, -58, -117, -51, 11, -12, -108, -118, -67, 72, 37, 108, 113, 90, 74, 89, 0, -106, -28, 37, 121, -106, -106, 9, -67, 115, -116, 69, 48, 19, -127, -65, 27, 36, -56, -95, 1, -116, 51, -1, 56, 65, 37, 118, -59, -117, -1, -99, 2, 13, 40, -89, 80, -35, 62, -79, -67, 75, 61, -11, 59, -98, -68, 77, 114, -68, -23, -97, 8, -1, -28, 67, -6, -39, 36, -99, -44, 24, 1, 94, -106, -88, 96, 73, -88, 113, 55, -15, -110, -49, 80, 99, 41, -71, -56, -84, -31, 22, 78, 68, -104, 63, 53, -73, 110, -1, -18, -48, -127, -4, 88, -24, -77, -100, 29, -105, -62, -79, 79, -32, -2, -100, -67, -68, 46, 61, -61, 112, -67, -26, -92, -48, -55, 104, -9, -114, -95, -59, -114, 71, 123, -117, 19, -33, -89, 62, 48, 108, -61, 27, 42, -128, -34, -55, 6, 91, 41, 32, -57, 99, -37, 21, -19, -19, 113, -69, 61, 96, 63, 111, 10, -111, -26, -20, 82, -69, 125, 30, 19, -88]
