+++ Commons/src/main/java/module-info.java
@@ -1,0 +1,7 @@
+module de.metallist.contractcollection.commons {
+    requires com.fasterxml.jackson.databind;
+    requires lombok;
+    requires org.slf4j;
+
+    exports de.metallistdev.contractcollection.commons;
+}
