+++ Frontends/javafx/src/main/java/de/metallistdev/contractcollection/javafx/ContractCollectionApplication.java
@@ -1,0 +1,24 @@
+package de.metallistdev.contractcollection.javafx;
+
+import javafx.application.Application;
+import javafx.fxml.FXMLLoader;
+import javafx.scene.Scene;
+import javafx.stage.Stage;
+
+import java.io.IOException;
+
+public class ContractCollectionApplication extends Application {
+
+    public static void main(String[] args) {
+        launch(args);
+    }
+
+    @Override
+    public void start(Stage primaryStage) throws IOException {
+        FXMLLoader fxmlLoader = new FXMLLoader(ContractCollectionApplication.class.getResource("login-view.fxml"));
+        Scene scene = new Scene(fxmlLoader.load(), 320, 240);
+        primaryStage.setTitle("Contract Collection");
+        primaryStage.setScene(scene);
+        primaryStage.show();
+    }
+}
