--- build.gradle
@@ -1,53 +1,0 @@
-plugins {
-    id 'org.springframework.boot' version '3.0.2'
-    id 'io.spring.dependency-management' version '1.0.11.RELEASE'
-    id 'java'
-}
-
-ext {
-    SpringVersion = '3.0.2'
-    TestNGVersion = '7.7.0'
-}
-
-group = 'de.metallist'
-version = '0.0.2-SNAPSHOT'
-sourceCompatibility = '17'
-targetCompatibility = '17'
-
-repositories {
-    mavenCentral()
-}
-
-dependencies {
-    /////////////////
-    // SPRING BOOT //
-    /////////////////
-
-    // Apache License 2.0
-    // https://mvnrepository.com/artifact/org.springframework.boot/spring-boot-starter
-    implementation group: 'org.springframework.boot', name: 'spring-boot-starter', version: SpringVersion
-
-    // Apache License 2.0
-    // https://mvnrepository.com/artifact/org.springframework.boot/spring-boot-starter-web
-    implementation group: 'org.springframework.boot', name: 'spring-boot-starter-web', version: SpringVersion
-
-    // Apache License 2.0
-    // https://mvnrepository.com/artifact/org.springframework.boot/spring-boot-starter-data-jpa
-    implementation group: 'org.springframework.boot', name: 'spring-boot-starter-data-jpa', version: SpringVersion
-
-    // Apache License 2.0
-    // https://mvnrepository.com/artifact/org.springframework.boot/spring-boot-devtools
-    developmentOnly group: 'org.springframework.boot', name: 'spring-boot-devtools', version: SpringVersion
-
-    // Apache License 2.0
-    // https://mvnrepository.com/artifact/org.springframework.boot/spring-boot-starter-test
-    testImplementation group: 'org.springframework.boot', name: 'spring-boot-starter-test', version: SpringVersion
-
-    // Apache License 2.0
-    // https://mvnrepository.com/artifact/org.testng/testng
-    testImplementation group: 'org.testng', name: 'testng', version: TestNGVersion
-}
-
-tasks.named('test') {
-    useTestNG()
-}
