--- contracts.json
@@ -1,40 +1,0 @@
-[ {
-  "id" : 1,
-  "category" : "insurance",
-  "name" : "health insurance",
-  "expenses" : 100.0,
-  "cycle" : 12,
-  "customerNr" : "12345",
-  "contractNr" : "67890",
-  "startDate" : "2022-01-01",
-  "contractPeriod" : 1,
-  "periodOfNotice" : 2,
-  "description" : "public health insurance - student tariff",
-  "documentPath" : "/home/user/example"
-}, {
-  "id" : 2,
-  "category" : "insurance",
-  "name" : "health insurance",
-  "expenses" : 100.0,
-  "cycle" : 12,
-  "customerNr" : "12345",
-  "contractNr" : "67890",
-  "startDate" : "2022-01-01",
-  "contractPeriod" : 1,
-  "periodOfNotice" : 2,
-  "description" : "public health insurance - student tariff",
-  "documentPath" : "/home/user/example"
-}, {
-  "id" : 3,
-  "category" : "insurance",
-  "name" : "health insurance",
-  "expenses" : 100.0,
-  "cycle" : 12,
-  "customerNr" : "12345",
-  "contractNr" : "67890",
-  "startDate" : "2022-01-01",
-  "contractPeriod" : 1,
-  "periodOfNotice" : 2,
-  "description" : "public health insurance - student tariff",
-  "documentPath" : "/home/user/example"
-} ]
