+++ settings.gradle.kts
@@ -1,0 +1,5 @@
+rootProject.name = "ContractCollection"
+
+include("Commons")
+include("Backend")
+include("Frontends:javafx")
