--- src/main/java/de/metallist/contractcollection/ContractCollectionApplication.java
@@ -1,13 +1,0 @@
-package de.metallist.contractcollection;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class ContractCollectionApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(ContractCollectionApplication.class, args);
-    }
-
-}
