+++ src/main/java/de/metallistdev/contractcollection/ContractCollectionApplication.java
@@ -1,0 +1,13 @@
+package de.metallistdev.contractcollection;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class ContractCollectionApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(ContractCollectionApplication.class, args);
+	}
+
+}
