--- src/test/java/de/metallist/contractcollection/ContractCollectionApplicationTests.java
@@ -1,16 +1,0 @@
-package de.metallist.contractcollection;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-import static org.testng.Assert.assertTrue;
-
-@SpringBootTest
-class ContractCollectionApplicationTests {
-
-    @Test
-    void contextLoads() {
-        assertTrue(true);
-    }
-
-}
