+++ src/test/java/de/metallistdev/contractcollection/ContractCollectionApplicationTests.java
@@ -1,0 +1,13 @@
+package de.metallistdev.contractcollection;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class ContractCollectionApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
